refactor: add pdf and tests
